# Reduced encodes