--- a/dist.ini
+++ b/dist.ini
@@ -17,6 +17,7 @@ issues = 1
 [Prereqs]
 perl = 5.010
 JSON::MaybeXS = 0
+Time::HiRes = 0
 
 [PruneFiles]
 match = ~$