--- a/lib/JSON/RPC2/TwoWay.pm
+++ b/lib/JSON/RPC2/TwoWay.pm
@@ -3,7 +3,7 @@ use 5.10.0;
 use strict;
 use warnings;
 
-our $VERSION = '0.07'; # VERSION
+our $VERSION = '0.08'; # VERSION
 
 # standard perl
 use Carp;