--- a/lib/JSON/RPC2/TwoWay/Connection.pm
+++ b/lib/JSON/RPC2/TwoWay/Connection.pm
@@ -4,7 +4,7 @@ use 5.10.0;
 use strict;
 use warnings;
 
-our $VERSION = '0.07'; # VERSION
+our $VERSION = '0.08'; # VERSION
 
 # standard perl
 use Carp;