Most of the payload is getting json encoded twice: once to generate the id and again for the json rpc message.

Here I take a very similar approach to `Mojo::IOLoop` for generating the id. 

I have removed the counter as monotonic time + `rand` should provide enough uniqueness. I am also using the `refaddr` of the argument most likely to change from call to call.

Let me know if you don't think this provides enough randomness.

## Benchmarks

I carried out some benchmarks by calling different versions of the functions with the write disabled to measure throughput:

* in the "best case" scenario the new version performs somewhat worse ~ 0.7x.
* in every other the new version performs between somewhat better ~ 1.4x and much better ~ 3x.

### Table

version | method | scenario | bytes | calls_per_sec | comparison
-- | -- | -- | -- | -- | --
new | callraw | best | 2 | 440528.63 | 70.93%
new | call | best | 2 | 328947.37 | 67.76%
new | callraw | average1 | 1011 | 289017.34 | 144.22%
new | call | average1 | 1011 | 243902.44 | 130.00%
new | callraw | average2 | 2011 | 226244.34 | 180.09%
new | call | average2 | 2011 | 204918.03 | 167.42%
new | callraw | worst | 900011 | 1196.17 | 310.77%
new | call | worst | 900011 | 1133.79 | 300.91%
old | callraw | best | 2 | 621118.01 | 140.99%
old | call | best | 2 | 485436.89 | 147.57%
old | callraw | average1 | 1011 | 200400.8 | 69.34%
old | call | average1 | 1011 | 187617.26 | 76.92%
old | callraw | average2 | 2011 | 125628.14 | 55.53%
old | call | average2 | 2011 | 122399.02 | 59.73%
old | callraw | worst | 900011 | 384.91 | 32.18%
old | call | worst | 900011 | 376.79 | 33.23%