@a6502 2023-07-17
My guess is that mono_time() . rand . refaddr($args) should produce enough randomness, at least on a per connection (or "channel") basis. If that hold up in practice you will see.